+++ crates/bcr-wdc-treasury-service/src/debit/service.rs
@@ -51,6 +51,15 @@ pub trait Repository: Clone + Send {
     async fn store_quote(&self, quote: MintQuote) -> Result<()>;
     async fn delete_quote(&self, qid: String) -> Result<()>;
     async fn list_quotes(&self) -> Result<Vec<MintQuote>>;
+    async fn store_onchain_melt(
+        &self,
+        quote_id: uuid::Uuid,
+        request: bcr_common::wire::melt::MeltQuoteOnchainRequest,
+    ) -> Result<()>;
+    async fn load_onchain_melt(
+        &self,
+        quote_id: uuid::Uuid,
+    ) -> Result<bcr_common::wire::melt::MeltQuoteOnchainRequest>;
 }
 
 #[derive(Clone)]
@@ -265,6 +274,8 @@ mod tests {
             async fn store_quote(&self, quote: MintQuote) -> Result<()>;
             async fn delete_quote(&self, qid: String) -> Result<()>;
             async fn list_quotes(&self) -> Result<Vec<MintQuote>>;
+            async fn store_onchain_melt(&self, quote_id: uuid::Uuid, request: bcr_common::wire::melt::MeltQuoteOnchainRequest) -> Result<()>;
+            async fn load_onchain_melt(&self, quote_id: uuid::Uuid) -> Result<bcr_common::wire::melt::MeltQuoteOnchainRequest>;
         }
     }
