+++ crates/bcr-wdc-treasury-service/src/lib.rs
@@ -8,7 +8,7 @@ use axum::{
 };
 use bcr_wdc_treasury_client::TreasuryClient;
 use bitcoin::secp256k1;
-use clwdr_client::SignatoryNatsClient;
+use clwdr_client::{ClowderNatsClient, SignatoryNatsClient};
 // ----- local modules
 mod admin;
 mod credit;
@@ -53,6 +53,7 @@ pub struct AppConfig {
     satonline_repo: persistence::surreal::ForeignOnlineConnectionConfig,
     satoffline_repo: persistence::surreal::ForeignOfflineConnectionConfig,
     clowder_url: reqwest::Url,
+    clwdr_nats_url: Option<reqwest::Url>,
     signer_url: reqwest::Url,
     sat_wallet: debit::CDKWalletConfig,
     wildcat: debit::WildcatClientConfig,
@@ -66,6 +67,7 @@ pub struct AppController {
     crsat: Arc<ProdCrsatService>,
     sat: Arc<ProdSatService>,
     signer: Arc<SignatoryNatsClient>,
+    clwdr_nats: Option<Arc<ClowderNatsClient>>,
     dev: Arc<devmode::Service>,
 }
 
@@ -75,6 +77,7 @@ impl AppController {
             credit_keys_url,
             cdk_mintd_url,
             clowder_url,
+            clwdr_nats_url,
             signer_url,
             credit_repo,
             debit_repo,
@@ -180,6 +183,16 @@ impl AppController {
             .await
             .expect("Failed to create signer");
 
+        let clwdr_nats = if let Some(url) = clwdr_nats_url {
+            Some(Arc::new(
+                ClowderNatsClient::new(url, false)
+                    .await
+                    .expect("Failed to create clowder nats client"),
+            ))
+        } else {
+            None
+        };
+
         let dev = devmode::Service {
             crkeys: bcr_common::client::keys::Client::new(credit_keys_url),
             dbmint: cdk::wallet::HttpClient::new(cdk_mintd_url),
@@ -190,6 +203,7 @@ impl AppController {
             crsat,
             sat,
             signer: Arc::new(signer),
+            clwdr_nats,
             dev: Arc::new(dev),
         }
     }
@@ -220,14 +234,8 @@ pub fn routes(app: AppController) -> Router {
             TreasuryClient::SATEXCHANGEOFFLINE_EP_V1,
             post(web::sat_offline_exchange),
         )
-        .route(
-            TreasuryClient::STOREONCHAINMELT_EP_V1,
-            post(web::store_onchain_melt),
-        )
-        .route(
-            TreasuryClient::LOADONCHAINMELT_EP_V1,
-            post(web::load_onchain_melt),
-        );
+        .route("/v1/melt/quote/onchain", post(web::melt_quote_onchain))
+        .route("/v1/melt/onchain", post(web::melt_onchain));
     let admin = Router::new()
         .route(
             "/v1/admin/treasury/debit/request_to_mint_from_ebill",
